requirements.txt
Django==5.0.1
djangorestframework==3.14.0
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL
django-allauth==0.57.0
django-cors-headers==4.3.1
python-decouple==3.8
drf-spectacular==0.27.0
Pillow>=10.3.0
requests==2.31.0
